refactor: 컴포넌트로 나누어 style 지정
