+++ src/pages/Home/components/Map/Map.module.scss
@@ -1,168 +1,11 @@
 @use "src/utils/styles/mediaQuery" as media;
 
-@keyframes show-out {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    visibility: visible;
-    opacity: 1;
-  }
-}
-
 .map {
   width: 100%;
   height: calc(100vh - 81px);
 
   @include media.media-breakpoint-down(mobile) {
     height: calc(100vh - 80px);
-  }
-}
-
-.bubble {
-  display: flex;
-  justify-content: space-evenly;
-  text-align: center;
-  border: 2px solid #ff7f23;
-  background-color: white;
-  border-radius: 32.5px;
-  min-width: 205px;
-  height: 65px;
-
-  &__text {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
-
-  img {
-    width: 58px;
-    height: 58px;
-    border-radius: 89.5px;
-  }
-
-  span {
-    padding-left: 15px;
-    font-size: 20px;
-    font-weight: 500;
-    text-overflow: ellipsis;
-  }
-
-  &--clicked {
-    display: flex;
-    justify-content: space-evenly;
-    border: 2px solid #ff7f23;
-    background-color: #ff7f23;
-    border-radius: 32.5px;
-    min-width: 205px;
-    height: 65px;
-
-    &__text {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-    }
-
-    img {
-      width: 58px;
-      height: 58px;
-      border-radius: 89.5px;
-    }
-
-    span {
-      padding-left: 15px;
-      font-size: 20px;
-      font-weight: 500;
-      text-overflow: ellipsis;
-    }
-  }
-}
-
-.triangle {
-  position: relative;
-  left: -125px;
-  top: 65px;
-
-  &__background {
-    border-color: #ff7f23 transparent transparent;
-    border-style: solid;
-    border-width: 15px 20px 0;
-  }
-
-  &__color {
-    border-color: white transparent transparent;
-    border-style: solid;
-    border-width: 15px 20px;
-    position: absolute;
-    top: -2px;
-    left: 0;
-
-    &--clicked {
-      border-color: #ff7f23 transparent transparent;
-    }
-  }
-}
-
-.top-options {
-  position: absolute;
-  top: 20px;
-  left: 50%;
-  transform: translate(-50%);
-  width: 80%;
-  z-index: 10;
-
-  &__search {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    height: 48px;
-    background: #ffffff;
-    box-shadow: 2px 3px 12px 1px rgb(0 0 0 / 10%);
-    opacity: 0.8;
-    border-radius: 100px;
-    text-decoration: none;
-    color: #666666;
-    padding-left: 18px;
-    padding-right: 15px;
-    margin-bottom: 10px;
-  }
-
-  &__list {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 4px;
-    justify-items: center;
-  }
-
-  &__text {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    box-sizing: border-box;
-    width: 90%;
-    height: 30px;
-    background: rgb(251 251 250 / 90%);
-    border: 1px solid #c4c4c4;
-    box-shadow: 2px 3px 12px 1px rgb(0 0 0 / 10%);
-    border-radius: 37px;
-    font-weight: 500;
-    font-size: 12px;
-    color: #595959;
-    cursor: pointer;
-
-    svg {
-      padding-right: 5px;
-    }
-
-    &--clicked {
-      border: 1px solid #ff7f23;
-      color: #ff7f23;
-
-      svg path {
-        fill: #ff7f23;
-      }
-    }
   }
 }
 
@@ -204,81 +47,5 @@
     background: #ffffff;
     box-shadow: 2px 3px 12px 1px rgb(0 0 0 / 10%);
     border-radius: 24px;
-  }
-}
-
-.button {
-  z-index: 10;
-  position: relative;
-  bottom: 50px;
-}
-
-.filter {
-  position: absolute;
-  right: 0;
-  bottom: -115px;
-  list-style: none;
-
-  &__button {
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    width: 142px;
-    height: 44px;
-    background-color: #222222;
-    border-radius: 22px;
-    border: none;
-    margin-top: 15px;
-    text-decoration: none;
-    font-size: 18px;
-    font-weight: 400;
-    color: #ffffff;
-    cursor: pointer;
-
-    &--clicked {
-      background-color: #f6bf54;
-    }
-  }
-
-  &__menu-list {
-    position: absolute;
-    bottom: 150px;
-    right: 50px;
-    display: flex;
-    flex-direction: column-reverse;
-    align-items: flex-end;
-
-    &--show {
-      width: 340px;
-
-      .slide-filter-list {
-        :nth-child(n) {
-          animation: show-out 0.2s ease-out normal forwards;
-        }
-      }
-    }
-
-    .slide-filter-list {
-      visibility: hidden;
-      justify-content: center;
-      display: flex;
-      flex-direction: column;
-      margin-top: 10px;
-
-      &__button {
-        width: 142px;
-        height: 40px;
-        border-radius: 20px;
-        box-sizing: border-box;
-        border: 1px solid #f6bf54;
-        background-color: #ffffff;
-        margin-top: 10px;
-        cursor: pointer;
-
-        &:hover {
-          text-decoration: underline;
-        }
-      }
-    }
   }
 }
