+++ src/pages/Home/components/Map/components/MobileOptions/MobileOptions.module.scss
@@ -1,0 +1,61 @@
+.top-options {
+  position: absolute;
+  top: 20px;
+  left: 50%;
+  transform: translate(-50%);
+  width: 80%;
+  z-index: 10;
+
+  &__search {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    height: 48px;
+    background: #ffffff;
+    box-shadow: 2px 3px 12px 1px rgb(0 0 0 / 10%);
+    opacity: 0.8;
+    border-radius: 100px;
+    text-decoration: none;
+    color: #666666;
+    padding-left: 18px;
+    padding-right: 15px;
+    margin-bottom: 10px;
+  }
+
+  &__list {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    gap: 4px;
+    justify-items: center;
+  }
+
+  &__text {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    box-sizing: border-box;
+    width: 90%;
+    height: 30px;
+    background: rgb(251 251 250 / 90%);
+    border: 1px solid #c4c4c4;
+    box-shadow: 2px 3px 12px 1px rgb(0 0 0 / 10%);
+    border-radius: 37px;
+    font-weight: 500;
+    font-size: 12px;
+    color: #595959;
+    cursor: pointer;
+
+    svg {
+      padding-right: 5px;
+    }
+
+    &--clicked {
+      border: 1px solid #ff7f23;
+      color: #ff7f23;
+
+      svg path {
+        fill: #ff7f23;
+      }
+    }
+  }
+}
